feat: create factory
